Make ViewControllers subscribe to and observe ViewModels
ViewControllers subscribe to ViewModels and also
use DisposeBag to dispose of observables.
