+++ StickySessions/StickySessions/ViewModel/ListSessions/ListSessionsViewModel.swift
@@ -11,46 +11,22 @@ import RxSwift
 
 class ListSessionsViewModel {
 
-    private let disposeBag = DisposeBag()
-
     let remoteAPI:RemoteAPI
     let sessionsRepository:SessionsRepositoryProtocol
 
-    let updateSessions: ([SessionViewModel]) -> Void
-    var sessionsViewModel:[SessionViewModel] = []
-
-
     // TODO: Replace instantiation with Dependency Injection
-    init(updateSessions: @escaping (([SessionViewModel]) -> Void)) {
-        self.updateSessions = updateSessions
+    init() {
         self.remoteAPI = AlamofireRemoteAPI()
         self.sessionsRepository = SessionsRemoteRepository(remoteAPI: AlamofireRemoteAPI())
     }
 
-    func fetchSessions() {
-        sessionsRepository.getSessions()
-            .observeOn(MainScheduler.instance)
-            .subscribe(
-                onNext: { sessions in
-                    let orderedSessions = sessions.sorted(
-                        by: { $0.timestamp._seconds > $1.timestamp._seconds }
-                    )
-
-                    self.sessionsViewModel = orderedSessions.compactMap {
-                        SessionViewModel(session: $0)
-                    }
-
-                    self.updateSessions(self.sessionsViewModel)
-                },
-                onError: { error in
-                    self.fail(errorMsg: error.localizedDescription)
-                })
-            .disposed(by: disposeBag)
-
-    }
-
-    func fail(errorMsg: String) {
-        print("Error: ", errorMsg)
+    func fetchSessions() -> Observable<[Session]> {
+        return sessionsRepository.getSessions()
+            .map({ sessions -> [Session] in
+                return sessions.sorted(
+                    by: { $0.timestamp._seconds > $1.timestamp._seconds }
+                )
+            })
     }
 
 }
